@@ -8,16 +8,18 @@
     "watch": "webpack serve --mode=development"
   },
   "dependencies": {
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/x-data-grid": "^8.14.0",
     "cal-heatmap": "^3.6.2",
     "dayjs": "^1.11.11",
     "flux": "2.1.1",
-    "griddle-react": "0.8.x",
     "jquery": "^3",
-    "react": "16.x",
-    "react-dom": "16.x",
-    "react-notification-system": "0.2.x",
-    "react-router": "3.x",
-    "react-skylight": "0.5.1"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-responsive-modal": "^7.1.0",
+    "react-router": "^7.9.4",
+    "react-router-dom": "^7.9.4"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.26.7",
@@ -25,7 +27,7 @@
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "babel-core": "6.x",
-    "babel-loader": "^9.2.1",
+    "babel-loader": "^10.0.0",
     "babel-preset-es2015": "6.x",
     "babel-preset-react": "6.x",
     "babel-preset-stage-1": "latest",
